# openapi-comparison/api-blueprint/generated-server/src/library_api/models/book.py
# coding: utf-8

"""
    Library Management API

    API quản lý thư viện — cho phép thực hiện CRUD trên sách và thành viên.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, StrictBool
from typing import Any, ClassVar, Dict, List, Optional, Union
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Book(BaseModel):
    """
    Book
    """ # noqa: E501
    id: float
    title: str
    author: str
    isbn: Optional[str] = None
    genre: str
    price: float
    published_year: Optional[float] = None
    description: Optional[str] = None
    available: StrictBool
    created_at: Optional[str] = None
    __properties: ClassVar[List[str]] = ["id", "title", "author", "isbn", "genre", "price", "published_year", "description", "available", "created_at"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Book from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Book from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "title": obj.get("title"),
            "author": obj.get("author"),
            "isbn": obj.get("isbn"),
            "genre": obj.get("genre"),
            "price": obj.get("price"),
            "published_year": obj.get("published_year"),
            "description": obj.get("description"),
            "available": obj.get("available"),
            "created_at": obj.get("created_at")
        })
        return _obj