# openapi-comparison/typespec/tsp-output/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Library Management API
  version: 0.0.0
tags:
  - name: Books
  - name: Members
paths:
  /books:
    get:
      operationId: Books_list
      description: Lấy danh sách tất cả sách. Hỗ trợ phân trang và tìm kiếm.
      parameters:
        - $ref: '#/components/parameters/PaginationParams.page'
        - $ref: '#/components/parameters/PaginationParams.limit'
        - $ref: '#/components/parameters/BookQueryParams.search'
        - $ref: '#/components/parameters/BookQueryParams.genre'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookListResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Books
    post:
      operationId: Books_create
      description: Tạo một bản ghi sách mới trong hệ thống.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
  /books/{id}:
    get:
      operationId: Books_read
      description: Lấy thông tin chi tiết của một cuốn sách theo ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Books
    put:
      operationId: Books_update
      description: Cập nhật toàn bộ thông tin của một cuốn sách.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
    delete:
      operationId: Books_remove
      description: Xoá vĩnh viễn một cuốn sách khỏi hệ thống.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Books
  /members:
    get:
      operationId: Members_list
      description: Lấy danh sách thành viên thư viện.
      parameters:
        - $ref: '#/components/parameters/PaginationParams.page'
        - $ref: '#/components/parameters/PaginationParams.limit'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Members
    post:
      operationId: Members_create
      description: Đăng ký thành viên mới.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Member'
                required:
                  - data
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Members
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
components:
  parameters:
    BookQueryParams.genre:
      name: genre
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Genre'
      explode: false
    BookQueryParams.search:
      name: search
      in: query
      required: false
      schema:
        type: string
      explode: false
    PaginationParams.limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 10
      explode: false
    PaginationParams.page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
      explode: false
  schemas:
    ApiError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
      description: Cấu trúc lỗi chuẩn
    Book:
      type: object
      required:
        - id
        - title
        - author
        - genre
        - price
        - available
      properties:
        id:
          type: integer
          format: int32
          description: ID tự tăng do server sinh ra
          readOnly: true
        title:
          type: string
          maxLength: 255
          description: Tiêu đề sách
        author:
          type: string
          maxLength: 255
          description: Tên tác giả
        isbn:
          type: string
          description: Mã ISBN (10 hoặc 13 chữ số)
        genre:
          allOf:
            - $ref: '#/components/schemas/Genre'
          description: Thể loại sách
        price:
          type: number
          format: float
          minimum: 0
          description: Giá bán (VNĐ)
        published_year:
          type: integer
          format: int32
          minimum: 1000
          maximum: 2100
          description: Năm xuất bản
        description:
          type: string
          maxLength: 1000
          description: Mô tả ngắn về nội dung sách
        available:
          type: boolean
          description: Sách còn hàng hay không
        created_at:
          type: string
          format: date-time
          description: Thời điểm tạo
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Thời điểm cập nhật lần cuối
          readOnly: true
      description: Đối tượng sách đầy đủ (bao gồm ID do server cấp)
    BookCreate:
      type: object
      required:
        - title
        - author
        - genre
        - price
      properties:
        title:
          type: string
          maxLength: 255
        author:
          type: string
          maxLength: 255
        isbn:
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        price:
          type: number
          format: float
          minimum: 0
        published_year:
          type: integer
          format: int32
          minimum: 1000
          maximum: 2100
        description:
          type: string
          maxLength: 1000
        available:
          type: boolean
          default: true
      description: Dữ liệu cần thiết để tạo hoặc cập nhật sách
    BookListResponse:
      type: object
      required:
        - total
        - page
        - limit
        - data
      properties:
        total:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/Book'
      description: Response wrapper cho danh sách sách
    BookResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Book'
      description: Response wrapper cho một cuốn sách
    Genre:
      type: string
      enum:
        - fiction
        - non-fiction
        - science
        - history
        - technology
        - children
    Member:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
        name:
          type: string
        email:
          type: string
        joined_date:
          type: string
          format: date
          readOnly: true
      description: Thành viên thư viện
    MemberCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
      description: Dữ liệu tạo thành viên mới
    MemberListResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/Member'
servers:
  - url: http://localhost:8003/api/v1
    description: Local development server
    variables: {}